+++ package.json
@@ -22,11 +22,11 @@
     "./packages/test/web-server",
     "./packages/core",
     "./packages/react-valibot",
-    "./packages/middleware",
     "./packages/redux-store",
     "./packages/styles",
     "./packages/support/cldr-data-downloader",
     "./packages/support/cldr-data",
+    "./packages/api-middleware",
     "./packages/api",
     "./packages/isomorphic-react",
     "./packages/isomorphic-react-dom",
