+++ packages/api-middleware/internal.js
@@ -1,3 +1,3 @@
 // This is required for Webpack 4 which does not support named exports.
 // eslint-disable-next-line no-undef
-module.exports = require('./dist/botframework-webchat-middleware.internal.js');
+module.exports = require('./dist/botframework-webchat-api-middleware.internal.js');
