+++ packages/api-middleware/legacy.js
@@ -1,3 +1,3 @@
 // This is required for Webpack 4 which does not support named exports.
 // eslint-disable-next-line no-undef
-module.exports = require('./dist/botframework-webchat-middleware.legacy.js');
+module.exports = require('./dist/botframework-webchat-api-middleware.legacy.js');
