+++ packages/api-middleware/package.json
@@ -1,38 +1,38 @@
 {
-  "name": "@msinternal/botframework-webchat-middleware",
+  "name": "@msinternal/botframework-webchat-api-middleware",
   "version": "0.0.0-0",
-  "description": "The botframework-webchat middleware package",
-  "main": "./dist/botframework-webchat-middleware.js",
-  "types": "./dist/botframework-webchat-middleware.d.ts",
+  "description": "botframework-webchat-api/middleware package",
+  "main": "./dist/botframework-webchat-api-middleware.js",
+  "types": "./dist/botframework-webchat-api-middleware.d.ts",
   "exports": {
     "./internal": {
       "import": {
-        "types": "./dist/botframework-webchat-middleware.internal.d.mts",
-        "default": "./dist/botframework-webchat-middleware.internal.mjs"
+        "types": "./dist/botframework-webchat-api-middleware.internal.d.mts",
+        "default": "./dist/botframework-webchat-api-middleware.internal.mjs"
       },
       "require": {
-        "types": "./dist/botframework-webchat-middleware.internal.d.ts",
-        "default": "./dist/botframework-webchat-middleware.internal.js"
+        "types": "./dist/botframework-webchat-api-middleware.internal.d.ts",
+        "default": "./dist/botframework-webchat-api-middleware.internal.js"
       }
     },
     "./legacy": {
       "import": {
-        "types": "./dist/botframework-webchat-middleware.legacy.d.mts",
-        "default": "./dist/botframework-webchat-middleware.legacy.mjs"
+        "types": "./dist/botframework-webchat-api-middleware.legacy.d.mts",
+        "default": "./dist/botframework-webchat-api-middleware.legacy.mjs"
       },
       "require": {
-        "types": "./dist/botframework-webchat-middleware.legacy.d.ts",
-        "default": "./dist/botframework-webchat-middleware.legacy.js"
+        "types": "./dist/botframework-webchat-api-middleware.legacy.d.ts",
+        "default": "./dist/botframework-webchat-api-middleware.legacy.js"
       }
     },
     ".": {
       "import": {
-        "types": "./dist/botframework-webchat-middleware.d.mts",
-        "default": "./dist/botframework-webchat-middleware.mjs"
+        "types": "./dist/botframework-webchat-api-middleware.d.mts",
+        "default": "./dist/botframework-webchat-api-middleware.mjs"
       },
       "require": {
-        "types": "./dist/botframework-webchat-middleware.d.ts",
-        "default": "./dist/botframework-webchat-middleware.js"
+        "types": "./dist/botframework-webchat-api-middleware.d.ts",
+        "default": "./dist/botframework-webchat-api-middleware.js"
       }
     },
     "./tsconfig.json": "./src/ts-config/config.json",
