+++ packages/api-middleware/src/private/templatePolyMiddleware.tsx
