+++ packages/api-middleware/src/types/LegacyComponentMiddleware.ts
