+++ packages/api-middleware/tsup.config.ts
@@ -4,9 +4,9 @@ import baseConfig from '../../tsup.base.config';
 const config: typeof baseConfig = {
   ...baseConfig,
   entry: {
-    'botframework-webchat-middleware': './src/index.ts',
-    'botframework-webchat-middleware.internal': './src/internal.ts',
-    'botframework-webchat-middleware.legacy': './src/legacy.ts'
+    'botframework-webchat-api-middleware': './src/index.ts',
+    'botframework-webchat-api-middleware.internal': './src/internal.ts',
+    'botframework-webchat-api-middleware.legacy': './src/legacy.ts'
   }
 };
