+++ packages/api/src/hooks/useRenderAttachment.ts
@@ -1,4 +1,4 @@
-import { type LegacyRenderAttachment } from '@msinternal/botframework-webchat-middleware/legacy';
+import { type LegacyRenderAttachment } from '@msinternal/botframework-webchat-api-middleware/legacy';
 import useWebChatAPIContext from './internal/useWebChatAPIContext';
 
 export default function useRenderAttachment(): LegacyRenderAttachment | undefined {
