+++ packages/component/.eslintrc.yml
@@ -9,5 +9,6 @@ env:
 rules:
   no-restricted-imports:
     - error
-    - message: 'To prevent duplicated import, @msinternal/botframework-webchat-middleware is only allowed to import in botframework-webchat-api.'
-      name: '@msinternal/botframework-webchat-middleware'
+    - message: 'To prevent duplicated import, @msinternal/botframework-webchat-api-* is only allowed to import in botframework-webchat-api.'
+      patterns:
+        - '@msinternal/botframework-webchat-api-*'
